+++ juniper_codegen/src/derive_enum.rs
