+++ juniper_tests/src/codegen/derive_enum.rs
