+++ juniper_tests/src/codegen/mod.rs
@@ -1,3 +1,3 @@
-mod enums;
-mod input_objects;
+mod derive_enum;
+mod derive_input_object;
 mod derive_object;
